chore: move studio utils to separate file
