+++ packages/react/__tests__/exports.ts
@@ -136,15 +136,15 @@ describe('@aws-amplify/ui-react/internal', () => {
           "IconVisibility",
           "IconVisibilityOff",
           "IconWarning",
-          "findChildOverrides",
-          "getOverrideProps",
-          "getOverridesFromVariants",
-          "mergeVariantsAndOverrides",
           "IconCheckCircleOutline",
           "IconFiberManualRecord",
           "IconHighlightOff",
           "PrimitiveCatalog",
           "createDataStorePredicate",
+          "findChildOverrides",
+          "getOverrideProps",
+          "getOverridesFromVariants",
+          "mergeVariantsAndOverrides",
           "useAuth",
           "useAuthSignOutAction",
           "useDataStoreBinding",
